==> Implementation/08. Filtering 2 - Function Word selection.ipynb <==
# 8. Filtering 2 - Function Word selection

#### This script loads the Function Words dataset and selects the most relevant ones, so as to not have too many sprseness in the data

### Imports

import pandas as pd
import numpy as np

### Definitions

all_fw_filename = 'data/FunctionWords-extracted.csv'

selected_fw_filename = 'data/FunctionWords-selected.txt'

### Load Bleached bigrams

df = pd.read_csv(all_fw_filename)

df.head()

### Get information on what is considered as 'low frequency'

total_count = np.asarray(df['FrequencyMale'] + df['FrequencyFemale'])

for i in [1, 10, 20, 50, 100, 1000, 10000]:
    print('Counts more than {0}: {1}'.format(i, np.sum(total_count > i)))

We will consider only those more than 50

### Remove columns with low frequency

df = df[total_count > 50]

total_count = total_count[total_count > 50]

### Calculate the entropy column

male_prob, female_prob = df['FrequencyMale']/total_count, df['FrequencyFemale']/total_count

male_prob.shape
df['Entropy'] = - ( male_prob * np.log2(male_prob) + female_prob * np.log2(female_prob) )

df.head()

### Take top 50 with least entropy (ie which discriminate the most)

df = df.sort_values('Entropy').reset_index(drop = True)

df = df.iloc[:50]

df.head()

### Save to new text file

with open(selected_fw_filename, 'w') as f:
    for word in df['Word']:
        f.write(word + '\n')
